Explicitly cast JAX random seed to int64.
PiperOrigin-RevId: 362006349

+++ dqn_zoo/c51/run_atari.py
@@ -84,7 +84,7 @@ def main(argv):
   logging.info('C51 on Atari on %s.', jax.lib.xla_bridge.get_backend().platform)
   random_state = np.random.RandomState(FLAGS.seed)
   rng_key = jax.random.PRNGKey(
-      random_state.randint(-sys.maxsize - 1, sys.maxsize + 1))
+      random_state.randint(-sys.maxsize - 1, sys.maxsize + 1, dtype=np.int64))
 
   if FLAGS.results_csv_path:
     writer = parts.CsvWriter(FLAGS.results_csv_path)
